Charge cell (#477)
* Fix vector serialization to be more generic (float, double, int support)
Charge cell test changes

* Charge cell implementation tweaks

* Fix linux build

+++ src/core/ecs/SignalExpression.cc
@@ -604,7 +604,7 @@ namespace ecs {
             //     src.expr,
             //     src.scope.String(),
             //     scope.String());
-            DebugAssertf(src.rootIndex >= 0 && src.rootIndex < src.nodeDebug.size(),
+            DebugAssertf(src.rootIndex >= 0 && (size_t)src.rootIndex < src.nodeDebug.size(),
                 "Saving invalid signal expression");
             dst = picojson::value(src.nodeDebug[src.rootIndex]);
         } else {
